requirements.txt
# Core Dependencies
flask==3.0.3
requests==2.31.0
pandas==2.0.3
groq==0.4.1
openai==1.6.1
langchain==0.1.0
langchain-groq==0.0.1

# Data Processing
openpyxl==3.1.2
python-dotenv==1.0.0
jinja2==3.1.6
WeasyPrint==59.0

# Web & API
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorama==0.4.6